Extend input controls to: keyboard, numpad, touch tap
So far the main input was pressing space, it's now the numpad 5 button
like on a Nokia 3310.  Not every one has a numpad though, so two
fallbacks are provided:

  * on desktops you can still press space
  * on mobile or other touch-only platforms tap the screen

Touch input is technically disallowed in the game rules but I really
want to support mobile and am hoping this will be considered fair use
because it's not swiping or tapping at elements on a screen, just
checking to see if it was tapped or not, i.e. simulating a button.