Print fast/slow to screen to debug touch controls
It's really hard to tell if the touch controls are working on mobile
because you die before you can touch anything, which is also strange
because it takes longer to die on desktop and it's supposedly random to
the same degree but printing this out will make it easier to check that
it works.  This text looks really horrible and totally the wrong size,
so I'll remove it right after this.