Limit updates to 15 TPS
The game goes really fast as it is right now and I haven't even added
any interesting challenges and can hardly test the collisions in time.

According to the game jam those who want to go the extra mile can limit
FPS (in our case TPS) to 15 which is similar to what the Nokia 3310 had
so I'm going to do that and probably remove the ticker delay that I had
in to artificially slow it down so far.  This might turn out to be a
really bad idea, we'll see.