Split game logic out into own module and restart when you die
Actually didn't want to reorganise all this code, it took about an hour
but I thought it would be worse if I carried on.  The part I like least
is the error returns for next screen but don't have a better idea right
now.  This could all be YAGNI if I don't end up having time to add more
screens, but at least it's a first example of not having most of the
game logic in a tangly mess in main.  So far...